% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_diffusion_pde.R
\name{calculate_coefficients}
\alias{calculate_coefficients}
\title{Set up PDE coefficients for fish abundance model}
\usage{
calculate_coefficients(pars, l_grid, l_interfaces)
}
\arguments{
\item{pars}{A list containing at least the model parameters \code{k}, \code{L_inf},
\code{d}, \code{m} and \code{r}}

\item{l_grid}{Size grid at cell centres}

\item{l_interfaces}{Size grid at cell interfaces}
}
\value{
A list containing \code{v} and \code{D} (at interfaces), and \code{mu} (at centres)
}
\description{
Set up PDE coefficients for fish abundance model
}
\keyword{internal}
