% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cod_CS_length_freq}
\alias{Cod_CS_length_freq}
\title{Example length frequency data for Cod in the Celtic Sea.}
\format{
A data frame with 377 rows and 3 columns. The columns are:
\describe{
\item{survey_date}{The within-year date of the survey (decimal year)}
\item{Length}{Length (cm)}
\item{count}{Number of fish observed with given length and annuli}
}
}
\usage{
Cod_CS_length_freq
}
\description{
Example length frequency data for Cod in the Celtic Sea.
}
\keyword{datasets}
