% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberDensity.R
\name{get_cohort_density}
\alias{get_cohort_density}
\title{Get cohort number density as a function of length and time}
\usage{
get_cohort_density(
  pars,
  l_max,
  Delta_l = 1,
  t_max = 10,
  Delta_t = 0.05,
  single_cohort = TRUE
)
}
\arguments{
\item{pars}{A list containing the model parameters: k, L_inf, d, m, r.}

\item{l_max}{The maximum size (cm)}

\item{Delta_l}{The size step size (cm).}

\item{t_max}{The maximum simulation time (years).}

\item{Delta_t}{The time step size (years). Default is 0.05.}

\item{single_cohort}{Logical. If TRUE (default) include only individuals born
within the first year. If FALSE, includes all individuals born between time
0 and time \code{t_max}.}
}
\value{
A matrix holding the number density u(t,l). Rows correspond to time
and columns to length.
}
\description{
This allows to follow a cohort size distribution through time.
}
\details{
Convolves the Green's function obtained with \code{get_greens_function()} with the
spawning density from \code{get_spawning_density()} to produce the number density
as a function of length and time.
}
