% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_diffusion_pde.R
\name{get_diffusion}
\alias{get_diffusion}
\title{Get diffusion rate}
\usage{
get_diffusion(pars, l_interfaces)
}
\arguments{
\item{pars}{A list containing at least the model parameter \code{d}}

\item{l_interfaces}{Size grid at cell interfaces}
}
\value{
A vector with the diffusion rate at the cell interfaces
}
\description{
The diffusion rate is
\deqn{D(l) = \frac{d l}{2}.}
}
