% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_diffusion_pde.R
\name{get_growth}
\alias{get_growth}
\title{Get growth rate}
\usage{
get_growth(pars, l_interfaces)
}
\arguments{
\item{pars}{A list containing at least the model parameters \code{k}, \code{L_inf} and
\code{r}}

\item{l_interfaces}{Size grid at cell interfaces}
}
\value{
A vector with the growth rate at the cell interfaces
}
\description{
The growth rate is the von Bertalanffy growth rate above a minimum size:
\deqn{g(l) = k(L_{inf}-l)}
and a constant growth rate r below that size.
}
