% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_diffusion_pde.R
\name{get_mortality}
\alias{get_mortality}
\title{Get mortality rate}
\usage{
get_mortality(pars, l_grid)
}
\arguments{
\item{pars}{A list containing at least the model parameter \code{m}}

\item{l_grid}{Size grid at cell centres}
}
\value{
A vector with the mortality rate at the cell centres
}
\description{
The mortality rate is
\deqn{\mu(l) = \frac{m}{l}.}
}
