% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthEstimation.R
\docType{package}
\name{growthEstimation-package}
\alias{growthEstimation}
\alias{growthEstimation-package}
\title{growthEstimation: Estimating growth rate parameters from age-at-size data}
\description{
Takes stochasticity in growth rates into account when estimating growth rates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gustavdelius/growthEstimation}
  \item \url{https://gustavdelius.github.io/growthEstimation}
  \item Report bugs at \url{https://github.com/gustavdelius/growthEstimation/issues}
}

}
\author{
\strong{Maintainer}: Jessica Westworth \email{jew20@st-andrews.ac.uk}

Authors:
\itemize{
  \item Gustav Delius \email{gustav.delius@york.ac.uk} (\href{https://orcid.org/0000-0003-4092-8228}{ORCID})
}

}
\keyword{internal}
