% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAge.R
\name{plot_age_likelihood}
\alias{plot_age_likelihood}
\title{Plot signed log-likelihood contributions from age_at_length with mean K lines}
\usage{
plot_age_likelihood(pars, age_at_length)
}
\arguments{
\item{pars}{A list of parameters.}

\item{age_at_length}{Data frame of age-at-length observations.}
}
\value{
A \code{ggplot2} object.
}
\description{
Creates a heatmap of the signed contribution of each cell to the NLL,
and overlays the mean observed and expected K values.
}
