% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity.R
\name{plot_density_3d}
\alias{plot_density_3d}
\title{Plot density function as an interactive 3D surface}
\usage{
plot_density_3d(u, Delta_l = 1, Delta_t = 0.05, l_min = 5, ...)
}
\arguments{
\item{u}{A numeric matrix containing the density (rows = time, columns =
size).}

\item{Delta_l}{Size step size used to compute \code{u}.}

\item{Delta_t}{Time step size used to compute \code{u}.}

\item{l_min}{Size at which the plot should start.}

\item{...}{Further arguments passed to \code{plotly::add_surface()}.}
}
\value{
A Plotly plot
}
\description{
Plot density function as an interactive 3D surface
}
\examples{
\dontrun{
pars <- list(k = 0.5, L_inf = 100, d = 1, m = 0.2)
u <- getDensity(pars, l_max = 100, Delta_l = 1, t_max = 10, Delta_t = 0.1)
plot_density_3d(u, Delta_l = 1, Delta_t = 0.1)
}
}
