% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAge.R
\name{plotAgeLikelihood}
\alias{plotAgeLikelihood}
\title{Plot the likelihood of the observed age at length data}
\usage{
plotAgeLikelihood(pars, age_at_length)
}
\arguments{
\item{pars}{A list of parameters}

\item{age_at_length}{Data frame of raw age-at-length observations; will be
preprocessed internally by \code{preprocess_length_at_age()}.}
}
\value{
A \code{ggplot2} object suitable for display in Shiny or saving.
}
\description{
Plot the likelihood of the observed age at length data
}
\examples{
# In practice provide a real `age_at_length` table for the species
# p <- plotAge(pars, age_at_length = df)
}
