% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_diffusion_pde.R
\name{setup_grid}
\alias{setup_grid}
\title{Set up spatial grid for PDE solver}
\usage{
setup_grid(N_l, Delta_l)
}
\arguments{
\item{N_l}{Number of size cells}

\item{Delta_l}{Size step size (cm)}
}
\value{
A list containing l_grid (cell centres) and l_interfaces
}
\description{
Set up spatial grid for PDE solver
}
\keyword{internal}
