% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_diffusion_pde.R
\name{solve_pde}
\alias{solve_pde}
\title{Numerically solve the PDE for fish abundance density}
\usage{
solve_pde(
  pars,
  u_initial,
  Delta_l = 1,
  t_max = 10,
  Delta_t = 0.05,
  min_observed_size = 0
)
}
\arguments{
\item{pars}{A list containing the model parameters: k, L_inf, d, m, r.}

\item{u_initial}{A numeric vector for the initial condition u(l, 0).}

\item{Delta_l}{The size step size (cm).}

\item{t_max}{The maximum simulation time.}

\item{Delta_t}{The time step size (years). Default is 0.05.}

\item{min_observed_size}{The minimum size observed in surveys (cm). Default is 0.}
}
\value{
A matrix where each column is the solution u(t,l) at a given time
step. Rows correspond to time and columns to length.
}
\description{
Implements an unconditionally stable finite volume scheme (implicit Euler)
with upwinding for advection.
}
