% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_diffusion_pde.R
\name{solve_pde_steady_state}
\alias{solve_pde_steady_state}
\title{Numerically solve the steady state of the PDE for fish abundance density}
\usage{
solve_pde_steady_state(pars, Delta_l = 1, l_max = 100)
}
\arguments{
\item{pars}{A list containing the model parameters: k, L_inf, d, m, r.}

\item{Delta_l}{The size step size (cm). Default is 1.}

\item{l_max}{The maximum size. Default is 100.}
}
\value{
A numeric vector containing the steady state solution u(l).
}
\description{
Finds the steady state solution by setting the time derivative to zero,
resulting in an ODE that is discretized using the same finite volume scheme
as solve_pde() and solved using Thomas's method.
}
