% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_diffusion_pde.R
\name{solve_thomas}
\alias{solve_thomas}
\title{Solve a tridiagonal system of equations using the Thomas algorithm}
\usage{
solve_thomas(a, b, c, d)
}
\arguments{
\item{a}{A numeric vector representing the sub-diagonal of A. Must have length N-1.}

\item{b}{A numeric vector representing the main diagonal of A. Must have length N.}

\item{c}{A numeric vector representing the super-diagonal of A. Must have length N-1.}

\item{d}{A numeric vector representing the right-hand side vector. Must have length N.}
}
\value{
A numeric vector containing the solution x.
}
\description{
This function solves the system Ax = d, where A is an NxN tridiagonal matrix.
}
