% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_diffusion_pde.R
\name{build_tridiagonal_system}
\alias{build_tridiagonal_system}
\title{Build tridiagonal system for PDE discretization}
\usage{
build_tridiagonal_system(N_l, Delta_l, Delta_t, coeffs)
}
\arguments{
\item{N_l}{Number of size cells}

\item{Delta_l}{Size step size}

\item{Delta_t}{Time step size (use 1.0 for steady state)}

\item{coeffs}{List of coefficients from calculate_coefficients()}
}
\value{
A list containing the tridiagonal matrix diagonals (a_, b_, c_)
}
\description{
Build tridiagonal system for PDE discretization
}
\keyword{internal}
