% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_diffusion_pde.R
\name{calculate_coefficients}
\alias{calculate_coefficients}
\title{Calculate PDE coefficients at grid points}
\usage{
calculate_coefficients(pars, l_grid, l_interfaces, steady_state = FALSE)
}
\arguments{
\item{pars}{A list containing model parameters: k, L_inf, d, m, r}

\item{l_grid}{Size grid at cell centres}

\item{l_interfaces}{Size grid at cell interfaces}

\item{steady_state}{If TRUE, use epsilon to avoid division by zero in mu}
}
\value{
A list containing v, D (at interfaces), and mu (at centres)
}
\description{
Calculate PDE coefficients at grid points
}
\keyword{internal}
