% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cod_CS_age_at_length}
\alias{Cod_CS_age_at_length}
\title{Example age_at_length data for Cod in the Celtic Sea.}
\format{
A data frame with 329 rows and 4 columns. The columns are:
\describe{
\item{survey_date}{The within-year date of the survey (decimal year)}
\item{length}{length (cm)}
\item{K}{Number of observed annuli}
\item{count}{Number of fish observed with given length and annuli}
}
}
\usage{
Cod_CS_age_at_length
}
\description{
Example age_at_length data for Cod in the Celtic Sea.
}
\keyword{datasets}
