% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cod_CS_length_freq}
\alias{Cod_CS_length_freq}
\title{Example length-frequency data for Cod in the Celtic Sea.}
\format{
A data frame with 103 rows and 2 columns. The columns are:
\describe{
\item{length}{length (cm)}
\item{count}{Number of fish observed with given length}
}
}
\usage{
Cod_CS_length_freq
}
\description{
Example length-frequency data for Cod in the Celtic Sea.
}
\keyword{datasets}
