% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateAge.R
\name{getLogLik}
\alias{getLogLik}
\title{Get log likelihood of age observations}
\usage{
getLogLik(pars, surveys, Delta_l = 1, Delta_t = 0.05)
}
\arguments{
\item{pars}{A list containing the model parameters: k, L_inf, d, m.}

\item{surveys}{A data frame with survey age-at-length observations with
columns \code{survey_date}, \code{Length}, \code{K}, and \code{count}.}

\item{dt}{Time step for the model simulation (years). Default is 0.05.}
}
\value{
A data frame with
}
\description{
Get log likelihood of age observations
}
