% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberDensity.R
\name{getNumberDensity}
\alias{getNumberDensity}
\title{Get number density as a function of length and time}
\usage{
getNumberDensity(pars, l_max, Delta_l = 1, t_max = 10, Delta_t = 0.05)
}
\arguments{
\item{pars}{A list containing the model parameters: k, L_inf, d, m.}

\item{l_max}{The maximum size}

\item{Delta_l}{The size step size (cm).}

\item{t_max}{The maximum simulation time.}

\item{Delta_t}{The time step size (years). Default is 0.05.}
}
\value{
A matrix holding the number density u(t,l). Rows correspond to time and columns to length.
}
\description{
Convolves the Green's function obtained with \code{getGreens()} with the
spawning density from \code{spawning_density()} to produce the number density
as a function of length and time.
}
