% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAge.R
\name{plot_log_likelihood}
\alias{plot_log_likelihood}
\title{Plot Signed Negative Log-Likelihood Contributions with Mean K Lines
Creates a heatmap of the signed contribution of each cell to the NLL,
and overlays the mean observed and expected K values.}
\usage{
plot_log_likelihood(contributions_df)
}
\arguments{
\item{contributions_df}{A data frame from calculate_and_aggregate_likelihood.}
}
\value{
A ggplot object.
}
\description{
Plot Signed Negative Log-Likelihood Contributions with Mean K Lines
Creates a heatmap of the signed contribution of each cell to the NLL,
and overlays the mean observed and expected K values.
}
